+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changes
 1.0rc4 (unreleased)
 -------------------
 
+- Introduce ``cone.app.browser.contextmenu.context_menu_group`` and
+  ``cone.app.browser.contextmenu.context_menu_item``.
+  [rnix, 2021-02-04]
+
 - Introduce ``cone.app.interfaces.IAuthenticator`` utility.
   [rnix, zworkb, 2021-02-02]